+++ framework/CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 Change Log
 
 - Enh #18236: Allow `yii\filters\RateLimiter` to accept a closure function for the `$user` property in order to assign values on runtime (nadar)
 - Bug #18233: Add PHP 8 support (samdark)
+- Bug #18229: Add flag for recognize SyBase databases on uses pdo_dblib (darkdef)
 
 
 2.0.37 August 07, 2020